adressing new comments
